Replace buggy FrameAndFill with FillHoles
`FrameAndFill()` gets buggy if an object passes through the seed point (currently the origin). Specifically, it will fill in the background along with the holes along the border. `FillHoles()` reliably fills all holes besides holes on the border. Since holes along the border turn out not to be important in the analysis, I would rather have the reliability of `FillHoles()`.